Activity 3
This is Activity 3 . Implemented "void" and "unvoid" functions in the models.py and handle them in views.py.

The void function will be called from views.py whenever a delete button is pressed against a row. This function will set the voided property to True, along with date_voided and void_reason fields.

Similarly, the unvoid function will reverse the void operation.